Meowpy_telegram/BotComponents/TemplateImageGen/requirements.txt
loguru
pillow
